\name{RepDbLite.Hsapiens.2012}
\docType{package}

\alias{RepDbLite.Hsapiens.2012-package}
\alias{RepDbLite.Hsapiens.2012}
\alias{RepDbLite.Hsapiens.2012}


\title{RepBase-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at RepBase on
  Mon Apr 18 18:25:39 2016 and based on the RepBase20_12
}

\author{.}



\examples{
## load the library
##library(RepDbLite.Hsapiens.2012)
## list the contents that are loaded into memory
ls('package:RepDbLite.Hsapiens.2012')
## show the db object that is loaded by calling it's name
RepDbLite.Hsapiens.2012

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
